Same code, using partials.
